cloud_storage_clients: rename error to s3_error
