+++ src/v/cloud_storage_clients/s3_client.cc
@@ -13,8 +13,8 @@
 #include "bytes/iobuf.h"
 #include "bytes/iobuf_istreambuf.h"
 #include "bytes/iobuf_parser.h"
-#include "cloud_storage_clients/error.h"
 #include "cloud_storage_clients/logger.h"
+#include "cloud_storage_clients/s3_error.h"
 #include "hashing/secure.h"
 #include "http/client.h"
 #include "net/connection.h"
