+++ src/v/cloud_storage_clients/s3_error.h
